requirements.txt
pandas
google-cloud-bigquery
fastapi
uvicorn